Squash all 4.2.2 changes to one
Prune selected call log filter options from overflow menu.

Bug: 7205096
Change-Id: Ib81e81195475e74f06ed7eda600b3f777a28ad70

Change dial pad to show chooser onStart instead of onCreate.

This fixes the problem where the chooser is not shown when the dialpad is
started and pushed to recents before a call comes in.  When the call comes
and dial pad is selected from recents, the dialpad is shown.  Instead, we
should be showing the chooser so the user has the option of getting back
into the call.

Bug: 5761141
Change-Id: I315117c6607dbb71e5359cf98785591d20c21881

Changed Quick Contacts details Button to ImageView

Removed unnecessary button to eliminate accessibility
confusion

Bug: 7434869
Change-Id: Ib503cde9a04566d8e086bb3671868f4a8059d0be

Fix blank edit view after orientation change.

If an orientation change occurs before the contact data loads from the
database, the code fails to reload the data because it is checking
against saved state (which is always present on an orientation change).
This fix changes the check to look for the actual data to be present
in the saved state.

Bug: 7122054
Change-Id: I4373eca14ba447abc57493cf4c5af6b978d8b7a2

Add content description for expand name fields button

Bug: 7497403
Change-Id: Ie0bf0b4ef66b0e1c58b718bcc0d2fda1a2478bcf

Add content description to More options button

Bug: 7434822
Change-Id: Iab67890ceb496590ffe452e380ac638310fc65dc

Fix incorrect tab highlight in QuickContactActivity.

Selected highlight is on wrong tab after rotation. Inherently, view pager
remembers the last selected tab.  But the highlight is done outside the
view pager and logic was missing to highlight properly when re-creating
from rotation.

Bug: 7206995
Change-Id: Ic016ddb743c323beba71ca95a0a1ae08566a7b1d

Minor changes to validateAction().

- Make error message more accurate.
- Change action validation to be inline instead of relying on for loop.

Bug: 7122054
Change-Id: I4d9e3e2a18659bbe5ccc0e355d1e7d97f5d38e0f

Changed content description for add contact and group buttons

Bug: 7416273
Change-Id: I485a8257e8afd0800ecc7913dc403170956c5644

Format phone number in QuickContactsActivity

Bug: 7230016

Change-Id: I0b28e131ee35d8948a574cab9b2a1c849edfc162

Keyboard nav fix for list in GroupDetailFragment

Bug: 7495057
Change-Id: I8847a6fc33f719e9ce828d26ba73367c13426483

Added ability to paste Arabic/Farsi digits into dialer

Bug: 6656969
Change-Id: I9a2c3a8ea5f5df0456c89d3af8b68f96d539dee9

Allow vanity numbers via paste/DIAL intent

Converts vanity numbers such as 1-800-FLOWERS to the numeric
equivalents when the number string is populated via a DIAL intent
or pasted from the clipboard. Does not remove valid formatting such
as dashes, periods, spaces, etc.

Bug: 7553166
Change-Id: I553c91628b8fd86ff38a8eeec7823ebdc96bfe6b

Change content description for phone right tab from...

"favorites" to "favorites and all contacts".

Bug 7589043

Change-Id: I3daececb58c7eae64738ed1ed017f85b6bcd15a5

Let talkback read People app name on launch

The root cause was that we were effectively doing setTitle(null) in onCreate.

We were replacing the people activity title with an intent extra, even if it's
not set.  Let's just remove this logic entirely because there's no point of
doing this, as the People activity doesn't show the app title.

This intent extra is not public API in ContactsContract, but it was in the
legacy API, so let's still keep ContactsRequest.set/getActivityTitle for now.
It still works for picker activities.

Bug 5508368

Change-Id: Id0ef790f9a7a721682de19d2c26d25e3622cee2d

Prevent going into single contact mode un-necessarily.

On a tablet, after creating a new contact, the contact list would enter
"single contact mode". This mode is intended for the case where you may
have an account filte set and you want to see a contact in a different
account.  Combined with a stale contact list, this caused the app to
go into single mode when it was not necessary.

This will fix the most common case where all accounts are shown and a
new contact is added. The new contact will be displayed in the normal list.

This will not fix the case where an account filter is active and a new
contact is created.  In this case, the new contact is still displayed in
single contact mode.

Also tested was the test cases in:
I233162572e25acb737c9eae787dfc146879a0dc2

Bug: 7621855
Change-Id: I6f323c4003677cec9db63cae347f4532071037b9

Don't show "set default" for directory contacts

Directory contacts don't allow setting (or clearing) of defaults
so don't populate those actions on the context menu.

Bug: 6563335
Change-Id: Ib96af2dd875c5fefe771360b05df39c01bec7c38

Cherry picking BitmapUtilTest fix.

Fixed on master in I85268c040498621a324c459ce237a51b30e7d3a4.

Bug: 7633565
Change-Id: I47fe32cdcc3780d42be99283f7b6e42b4e4ab940

Fixing unit test to not rely on html markup.

These unit tests broke due to a change in the framework Html class and markup.
https://googleplex-android-review.googlesource.com/#/c/236067/

Since the unit tests have nothing to do with html, it is better to
change the unit tests so they do not rely on the html markup.

Bug: 7352945
Change-Id: I3340575c08687fe481e91b3897037960911b1bbb

Fix Derp

Change-Id: I593ca892153644701870fe05e202c36878b8f7f6

Fix onStart()

Change-Id: Ic1ff2cdea34c09fcffefd3b81a871f519196f405