Add Call statistics to Contacts update
Available from the local menu in the Phone application (except from
the Dialer Fragment).

Call-type filter can be selected in the actionbar.
If all call types are selected the incoming/outgoing rates show up in
blue/green respectively.
Tapping on an item starts a new activity that shows all call types with
details for the given contact or number.

This commit optimizes and changes some parts of the initial commit.

- Cleanup, avoid code duplication, avoid broken call log for
     non-English locales
- Add German translation, cleanup time formatting
- Fix counting issues in header when not displaying all calls
- Improve layout and duration formatting
- Tweak details activity layout
- Show number of incoming/outgoing calls in details activity.
- Allow sorting by date, move filter reset out of date picker,
     prevent invalid date ranges, cleanup
- Reduce code duplication
- Fix issues in number handling
- Improve call lookup
- move https://github.com/SlimRoms/packages_apps_Contacts/commit/6f82496f889bb1b9f025c9d683fcbc2274ee39bc
  to calllog/CallLogAdapterHelper due off code cleanup in CallLogAdapter

Change-Id: I94cd49e167b82bc7eaeab9225ed96a273a33b18c